DataDiffFull/requirements.txt
beautifulsoup4==4.6.0
lxml
mock==1.0.1
pytest==3.0.7
pytest-pep8==1.0.6
python-dateutil
pytz==2016.7
requests[security]
retrying==1.3.3
robotframework==3.0
robotframework-difflibrary==0.0.1
robotframework-requests==0.4.7
robotframework-selenium2library==1.8.0
sh==1.12.14
jinja2==2.9.6
six
psutil
